Transfer responsibility of checking input to 'sleepsort'
Rational: Python, being a dynamic language may present pass a
potentially mixed-type iterable to the function 'sleepsort.' The
precondition for a successful run is that the client supply an
interable whose elements can be safely converted to int. An
exception will occur if it does not.

In an earlier version, I had the convesion during the configuration
of the threads. I decided that this was not the place to be allowing
such an exception to occur and decided to guard against it. This
cleans up the code a bit, essentially checking the input.